+++ cvat-ui/src/utils/clamp-opacity.ts
